requirements.txt
aiogram==3.10.0
python-dotenv==1.0.0
torch==2.4.1
openai-whisper==20231117
g4f==0.3.9.7
sqlalchemy==2.0.36
aiosqlite==0.20.0
openai==1.50.2
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.154.0
cryptography==44.0.0
pytz==2024.2
python-dateutil==2.9.0.post0
ffmpeg-python==0.2.0
aiohttp==3.9.5